Unify the three APK install paths behind one installApp
Local drag-&-drop, catalog download, and the provisioning worker each
had their own install code (sync-push vs installFromUrl vs -r -d, three
progress idioms, two Success/Failure checks). Collapse them into a single
installApp(adb, source, opts) in src/lib/adb/install.ts where source is a
local File or a URL list (+ optional sha256), sharing one pm install and
temp-file cleanup. online-install.ts keeps only the device-fetch
primitives (deviceFetchText/detectDownloader); the provisioning broker
routes its installFromUrl RPC through the unified installer. InstallStage
now has one definition, imported by the runtime types.
